Refactor Falcon-Perception model to streamline perception generation and update demo notebook
- Introduced `generate_perception` method in the Model class for improved object detection and segmentation, consolidating previous functionality.
- Removed the `generate_perception` function from the processing module to centralize perception logic within the model.
- Updated the demo notebook to reflect changes in the perception generation process, enhancing clarity and usability.
- Revised README to include updated model details and usage examples for both Falcon-Perception and Falcon-Perception-300M models.
