+++ mlx_vlm/models/falcon_perception/__init__.py
@@ -2,4 +2,4 @@ from . import processing_falcon_perception  # noqa: F401
 from .config import ModelConfig, TextConfig, VisionConfig
 from .falcon_perception import Model, VisionModel
 from .language import LanguageModel
-from .processing_falcon_perception import generate_perception, plot_detections
+from .processing_falcon_perception import plot_detections
