+++ mlx_vlm/models/falcon_perception/processing_falcon_perception.py
@@ -274,170 +274,6 @@ class FalconPerceptionProcessor:
         return expanded
 
 
-def generate_perception(
-    model,
-    processor,
-    *,
-    image,
-    query: str,
-    max_new_tokens: int = 512,
-    temperature: float = 0.0,
-    segm_threshold: float = 0.5,
-) -> List[Dict]:
-    """Run Falcon Perception detection with proper coord/size/seg decode loop.
-
-    Returns list of detections, each with:
-      - 'xy': dict with 'x', 'y' (center coords, normalized 0-1)
-      - 'hw': dict with 'h', 'w' (size, as fraction of image)
-      - 'mask': (H, W) binary mx.array segmentation mask (if seg token decoded)
-    """
-    import mlx.core as mx
-
-    from mlx_vlm.utils import load_image as _load_image
-
-    if not isinstance(image, Image.Image):
-        image = _load_image(image)
-    image = image.convert("RGB")
-    orig_w, orig_h = image.size
-
-    from ..base import to_mlx
-
-    result = processor(text=[query], images=[image], padding=False)
-    result = to_mlx(result)
-    input_ids = result["input_ids"]
-    pixel_values = result["pixel_values"]
-    image_grid_hw = result.get("image_grid_hw", None)
-
-    config = model.config
-    coord_token_id = config.coord_token_id
-    size_token_id = config.size_token_id
-    seg_token_id = config.seg_token_id
-    eos_id = config.eos_id
-
-    # Get grid dimensions for segmentation
-    if image_grid_hw is not None:
-        grid_h = int(image_grid_hw[0, 0].item())
-        grid_w = int(image_grid_hw[0, 1].item())
-    else:
-        ps = config.vision_config.spatial_patch_size
-        grid_h = orig_h // ps
-        grid_w = orig_w // ps
-
-    from ..cache import make_prompt_cache
-
-    cache = make_prompt_cache(model)
-
-    # Prefill: run full model with image
-    logits_out = model(
-        input_ids,
-        pixel_values=pixel_values,
-        cache=cache,
-        image_grid_hw=image_grid_hw,
-    )
-    logits = logits_out.logits
-
-    # Compute segmentation features from prefill hidden states (via AnyUp if available)
-    segm_features = None
-    if hasattr(model, "conv_segm"):
-        prefill_hidden = model.last_hidden_state
-        if prefill_hidden is not None:
-            segm_features = model.compute_segm_features(
-                prefill_hidden, input_ids, pixel_values, grid_h, grid_w
-            )
-            mx.eval(segm_features)
-
-    # After prefill, clear _position_ids so decode falls back to _rope_delta.
-    # _position_ids is sized for prefill only; indexing it at decode offsets fails.
-    lm = model.language_model
-    lm._position_ids = None
-    lm._pos_hw = None
-    # _rope_delta and _full_attn_mask are kept for correct decode positions.
-
-    embed_fn = lm.model.embed_tokens
-
-    detections = []
-    current_det = {}
-    coord_xy = mx.zeros((1, 2))
-    size_hw_val = mx.zeros((1, 2))
-
-    for step in range(max_new_tokens):
-        if temperature == 0.0:
-            token = mx.argmax(logits[:, -1, :], axis=-1)
-        else:
-            logits_scaled = logits[:, -1, :] / temperature
-            token = mx.random.categorical(logits_scaled)
-
-        token_id = token.item()
-
-        if token_id == eos_id:
-            break
-
-        token_2d = token.reshape(1, 1)
-
-        # Decode coord/size/seg from the hidden state of the PREVIOUS step
-        h_state = model.last_hidden_state
-        h_last = h_state[:, -1:, :] if h_state is not None else None
-
-        if token_id == coord_token_id and h_last is not None:
-            # Commit previous detection before starting a new one
-            if "xy" in current_det and "hw" in current_det:
-                detections.append(current_det)
-                current_det = {}
-
-            coord_logits = model.decode_coords(h_last.reshape(-1, h_last.shape[-1]))
-            num_bins = coord_logits.shape[-1]
-            pred_bins = mx.argmax(coord_logits, axis=-1)
-            pred_x = pred_bins[0, 0].item() / (num_bins - 1)
-            pred_y = pred_bins[0, 1].item() / (num_bins - 1)
-            coord_xy = mx.array([[pred_x, pred_y]])
-            current_det["xy"] = {"x": pred_x, "y": pred_y}
-
-        elif token_id == size_token_id and h_last is not None:
-            size_logits = model.decode_sizes(h_last.reshape(-1, h_last.shape[-1]))
-            hw_pred = model.process_sizes(size_logits)
-            pred_h = hw_pred[0, 0].item()
-            pred_w = hw_pred[0, 1].item()
-            size_hw_val = mx.array([[pred_h, pred_w]])
-            current_det["hw"] = {"h": pred_h, "w": pred_w}
-
-        elif token_id == seg_token_id and h_last is not None:
-            if segm_features is not None:
-                seg_hidden = h_last.reshape(-1)  # (D,)
-                mask = model.decode_segm_mask(
-                    seg_hidden, segm_features, orig_h, orig_w, segm_threshold
-                )
-                mx.eval(mask)
-                current_det["mask"] = mask
-            if "xy" in current_det and "hw" in current_det:
-                detections.append(current_det)
-            current_det = {}
-
-        # Decode step: embed token, apply coord/size encoding, run LM
-        embeds = embed_fn(token_2d)
-        embeds = model.encode_coords_into_embeds(
-            embeds,
-            token_2d,
-            coord_xy if token_id == coord_token_id else None,
-        )
-        embeds = model.encode_sizes_into_embeds(
-            embeds,
-            token_2d,
-            size_hw_val if token_id == size_token_id else None,
-        )
-        logits_out = lm(
-            token_2d,
-            inputs_embeds=embeds,
-            cache=cache,
-        )
-        logits = logits_out.logits
-        mx.eval(logits)
-
-    if "xy" in current_det and "hw" in current_det:
-        detections.append(current_det)
-
-    return detections
-
-
 def plot_detections(image, detections, save_path="perception_output.png"):
     """Plot bounding box detections and segmentation masks on the image and save."""
     import matplotlib
@@ -521,6 +357,5 @@ install_auto_processor_patch("falcon_perception", FalconPerceptionProcessor)
 
 __all__ = [
     "FalconPerceptionProcessor",
-    "generate_perception",
     "plot_detections",
 ]
